+++ server_mode/static/src/css/ribbon.css
@@ -2,7 +2,7 @@
     width: 200px;
     background: #e43;
     position: absolute;
-    top: 25px;
+    top: 50px;
     left: -50px;
     text-align: center;
     line-height: 50px;
@@ -13,7 +13,7 @@
     -moz-transform: rotate(-45deg);
     -o-transform: rotate(-45deg);
     transform: rotate(-45deg);
-    z-index: 10000;
+    z-index: 1000;
     position: fixed;
     box-shadow: 0 0 3px rgba(0,0,0,.3);
     background: rgba(255, 0, 0, .6);;
