+++ support_branding_adhoc/__openerp__.py
@@ -39,7 +39,6 @@ Set parameter for adhoc support branding
     ],
     'data': [
         'ir_config_parameter.xml',
-        'support.xml',
     ],
     'demo': [
     ],
