+++ web_support_client/models/__init__.py
@@ -4,4 +4,5 @@
 # directory
 ##############################################################################
 from . import support
+from . import res_users
 # vim:expandtab:smartindent:tabstop=4:softtabstop=4:shiftwidth=4:
