+++ web_support_client_doc/__openerp__.py
@@ -27,6 +27,7 @@
     'description': """
 Web Support Client with Website Documentation Integration
 =========================================================
+TODO: check if we need this module or we remove it
     """,
     'author':  'ADHOC SA',
     'website': 'www.adhoc.com.ar',
