+++ web_support_server/models/__init__.py
@@ -4,4 +4,5 @@
 # directory
 ##############################################################################
 from . import res_users
+from . import res_partner
 # vim:expandtab:smartindent:tabstop=4:softtabstop=4:shiftwidth=4:
