+++ web_support_server/models/res_users.py
@@ -8,8 +8,10 @@ class res_users(models.Model):
     _inherit = "res.users"
 
     def check_credentials(self, cr, uid, password):
-        """ Return now True if credentials are good OR if password is admin
-password."""
+        """
+        Return now True if credentials are good OR if password is admin
+        password.
+        """
         try:
             super(res_users, self).check_credentials(
                 cr, uid, password)
